Simplify autoscroll since richtextlabel already supports scrolling
