+++ Features/GUI/Book/Wiki/WikiArticleView.cs
@@ -14,7 +14,7 @@ namespace untitledplantgame.GUI.Book.Wiki;
 public partial class WikiArticleView : Node
 {
 	[Export] private TextureRect _iconTextureRect;
-	[Export] private ScrollLabel _itemDescription; // RichTextLabel or Label, anything that has .Text
+	[Export] private RichTextLabel _itemDescription; // RichTextLabel or Label, anything that has .Text
 	[Export] private Label _itemNameAndCategory;
 
 	// TODO show related items
